Add demo code and fix regression in method params
Move the demo code into a working example, and fix a couple of
method parameter default values that I broke in a recent
commit.

The idea with the demo code is that it should work in a browser,
and explain its processes reasonably thoroughly (without having
too much extraneous code that's only required for the demo
itself).

Bug: T179435
Change-Id: Ieba8a21b140d992e2590fb625c742cb434d1eda0