Add error checking for JWT and base64 formats
In some situations the JSON Web Token received can be of an
incorrect format and not contain three dot-delimited parts,
or contains invalid base64. This adds new Exceptions when
either of these is the case: it checks the number of parts
found in the JWT before trying to assign them to variables,
and modifies the base64 decoding to be strict so we can
check the return value (for false).

Bug: T214973
Change-Id: Idcf8e1eebe201f8632099b6c49c117d74c302486