Support authentication only sessions
The client so far has exclusively used the authorize endpoint to
initiate sessions, but mediawiki supports a second mode. The existing
mode always prompts the user to re-verify the authorize request,
but in cases where only authentication is being done we can use the
authenticate endpoint and skip the tedium of having the user re-accept
on every usage.

Adds a new configuration value: authenticateOnly. This value defaults
to false, which retains the existing functionality. Passing true
allows attempting auth only. There doesn't appear to be any downside
to always enabling authenticate only, the backend will recognize
requests that are not allowed to use authenticate-only and degrade
them into authorize requests. But still, the historical behaviour was
mantained to avoid opportunity of breaking unknown systems.

Change-Id: I79496d5779225bafa8f8169a7dd88602197e845d